Refactor: move all widgets to widget package
