+++ lib/widget/Audio_Library_App.dart
@@ -1,0 +1,19 @@
+import 'package:flutter/material.dart';
+
+import 'Audio_library_homepage.dart';
+
+class AudioLibraryApp extends StatelessWidget {
+  const AudioLibraryApp({Key? key}) : super(key: key);
+
+  @override
+  Widget build(BuildContext context) {
+    print("Hell!!!");
+    return MaterialApp(
+      title: 'Audio Library App',
+      theme: ThemeData(
+        primarySwatch: Colors.blue,
+      ),
+      home: AudioLibraryHomePage(key: UniqueKey()),
+    );
+  }
+}
